Modified BubbleSort.
The provious code sorted an array in a bubble-down approach, where the maximum element was moved to the last position at each iteration.
The modified code does in a bubble-up approach, where the minimum element is moved to the first position at each iteration.
I made this change, considering what "Bubble Sort" was named after.